Factor out other test
